Clarify meaning of 'has potential to develop into'.
There has been confusion about the meaning of RO:0002387 ('has potential
to develop into'), especially regarding the implication of that relation
on taxon constraints (the main question being: if 'X RO:0002387 some Y'
and Y is restricted to a given taxon, should that imply that X is also
restricted to that same taxon?).

We add a comment to tentatively clarify the intended meaning of that
relation, as well as a property chain over 'in taxon' (RO:0002162) to
formally answer the question about its implication on taxon constraints.
That property chain is currently being defined in Uberon, from where it
should be removed once the present commit makes its way into a RO
release.

closes #810