Do not define property chains over "grouping" properties.
Grouping properties are only intended to provide a nice hierarchy for
human users to browse, and should not be used in any kind of logical
axioms.

So we replace the previously introduced property chains over 'in taxon'
and (1) 'mereotopologicall related to', (2) 'developmentally related
to', and (3) 'relation between physical entity and a process or stage'
by several chains over some of their subproperties.