Add/streamline property chains over 'in taxon' (RO:0002162). (#832)
* Add/streamline property chains over 'in taxon' (RO:0002162).

This commit adds several property chains of the form

  X o 'in taxon' SubPropertyOf: 'in taxon'

where X represents the following properties:

* 'mereotologically related to' (RO:0002323),
* 'developmentally related to' (RO:0002324),
* 'relation between physical entity and a process or stage' (RO:0002487),
* 'innervated_by' (RO:0002005),
* 'innvervates' (RO:0002134),
* 'produced by' (RO:0003001),
* 'produces' (RO:0003000).

Existing property chains over properties that are subproperties of those
listed above (e.g. 'part of', BFO:0000050) are removed since they are no
longer necessary.

closes #831

* Do not define property chains over "grouping" properties.

Grouping properties are only intended to provide a nice hierarchy for
human users to browse, and should not be used in any kind of logical
axioms.

So we replace the previously introduced property chains over 'in taxon'
and (1) 'mereotopologicall related to', (2) 'developmentally related
to', and (3) 'relation between physical entity and a process or stage'
by several chains over some of their subproperties.

* Add property chain over 'innervated by' and 'in taxon'.

This is another chain defined by Uberon which we are moving to RO.